refactor: Activity 이름 변경, 중복 라이브러리 제거
