+++ gradle/libs.versions.toml
@@ -18,7 +18,6 @@ naverMapCompose = "1.7.2"
 navigationCompose = "2.8.2"
 playServicesLocation = "21.3.0"
 retrofit = "2.11.0"
-navigation = "2.8.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -47,7 +46,6 @@ androidx-material3 = { group = "androidx.compose.material3", name = "material3",
 naver-map-compose = { module = "io.github.fornewid:naver-map-compose", version.ref = "naverMapCompose" }
 play-services-location = { module = "com.google.android.gms:play-services-location", version.ref = "playServicesLocation" }
 retrofit = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit" }
-androidx-navigation = { group = "androidx.navigation", name = "navigation-compose", version.ref = "navigation"}
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
